Create Node namespace and load it into Alces
A basic node namespace has been created with the code to load the
config and answer hashes. The Node namespace is then loaded into the
Alces namespace as an array of all the nodes.

To do this, a new NodeattrInterface method was required to obtain a
complete list of all the nodes. Also the data structure used by
HashMerger has been moved to Constants

A merge method has been added to the HashMergers.rb module which can
create Answer and Config hashes. This makes it easier to create them
all at once. It uses a regular OpenStruct as it does not need to be
recursive nor does it need to template itself.