Switch alces input to block in HashMerger
The only reason for the alces namespace to be passed into the Hash
object is so it can be used as the templater. However this creates
coupling between the two objects.

It also isn't immediately clear why the alces namespace is being
included and then deleted from the hash.

Instead this action is better suited as a block which can be yeiled
to. The next commit will switch MetalRecursiveOpenStruct to yeild to
the block instead of calling a magic method on `alces`