Use a templater block in MetalRecursiveOpenStruct
Instead of passing in the alces object which created coupling; the
previous commit made it that a templater block passed instead.

This commit switches MetalRecursiveOpenStruct to use the block and
thus it does not need to know how to template anything.

Also removed the answer method from the Alces namespace as it is
not required.