Allow Node namespace to retrieve config values
NOTE: rubocop was ran on this commit which has lead to the large number
of files changed.

The config hash is loaded into the Node namespace using the HashMerger.
This commit adds additional code to pass in the templating block to use
when rendering the config values.

Whilst the config values can be rendered against the Alces namespace,
they currently can not reference other config values. This is because
a node can not be in scope ATM.

Therefore only simple config values (aka no references to other configs)
can be retreived at this point in time.