Create the dynamic namespace for the Node
Inorder for a node's config parameters to references itself, the Node
namespace needs to be loaded into the Alces namespace. This will allow
the config's to reference `alces.node.config` to retrieve other config
parameters.

This is done by keeping track of the dynamic namespace inside of the
Alces namespace. The dynamic namespace is also used to track the number
of recursion to prevent an infinite loop.