Fix bug where OpenStruct defines methods implicitly
By default OpenStruct will define a method dynamically when called.
This means subsequent calls of the same method do not need to go
through method_missing. However in the process it skips the renderer
and thus a config parameter can only be rendered once.

The fix for this, is not to inherit from OpenStruct. Instead the
behaviour of MetalRecursiveOpenStruct will be completely defined by
us. This will prevent the render being skipped silently.