Various changes and bug fixes
This commit makes various small changes to make the new templater work
with the existing repo.

It adds a benchmarking script for the new templater.

It makes the MetalRecursiveOpenStruct enumerable and return a nil if
a parameter is missing instead of erroring. This will need to be
combined with nil detection at a future date.

To prevent name configs in the Alces namespace, the metalware config
has been renamed.

Also a temporary fix has been applied to the Alces namespace so
templating parameters can be accessed in the same manor as 2017.1.0.
This is intended to be removed at a later date.

The name of a node has been made public and the config and answers for
a node are loaded into the dynamic namespace along with the Node
namespace itself

Rubocop has also been ran to make formatting fixes