Move permanet Alces Namespace methods to mixin
The alces namespace now has a dual role of being the templater and
binding to the templater.

Because of this, the alces.rb file itself will be kept for templater
related code. As the dynamic namespace requires inputs from the
templater, its code will be located here as well.

All the remaining permanent Namespaces that exist in the Alces namespace
shall be moved to a mixin module. This makes it clearer what each set of
code does.

Also removed recursion_count from the Alces namespace as it is nolonger
used.