Remove metalware config as input to Node namespace
The metalware config can be retrieved from the Alces namespace and
does not need to be passed seperately. This does require the used of
calling a private method, however this is to prevent exposing the
config to the user.

However doing it this way ensures the same config is used Namespace
wide.