Base Nodes namespace of Array and implement find
The Nodes namespace is better implemented as an array as instead of
a wrapper around one.

Once the array of Node namespaces is made, each node it added as a
method onto the Nodes namespace. This allows nodes to be accessed
by name

Once it is created, it can not be modified as nodes should not be
deleted or added into a existing Alces namespace.

Also added additional bug fixes where the Nodes namespace wasn't being
cached and respond_to? in Alces was missing an input.

The MissingParameter error in Templating::Rendered has been commente
out as it obfuscates the errors in development. A test has been skipped
as it is failing because of this. However as it is for the old
templater, it probably wont be fixed.