Create base class for hash merger namespaces
The Node, Group and Domain namespaces will all be based on HashMerger.
To facilitate this, a HashMegerNamespace has been created to faciliatate
this.

The HashMergers module contains the code on how to perfrom the
HashMerger for Config/Answer's. The HashMergerNamespace however contains
the code on how to link it to the Domain, Group and Node. It also
contains the code on how the templating should be performed.

Also created the Group namespace using the base class