Remove Nodes namespaces and replace with MetalArray
There are going to be a few Node, Group Namespace arrays which need
to be accessable using method calls.

To do this, a MetalArray has been created that defines methods according
to the name of each of item within the array.

This means the Nodes namespace is no longer required as it can be
replaced with a builder method for the MetalArray.

MetalArray will be used for alces.groups and alces.group.nodes