Implement the index method in Node namespace
The index for a node is defined as the position within it's primary
group as returned by nodeattr.

This means the Node namespace must search it's primary Group namespace
for it's position. This means the nodes in a Group has been implemented
as well.

Instead of recreating the Node namespaces, they are retrieved from the
Alces namespace so they caching can be shared.