Add caching to MetalRecursiveOpenStruct
Instead of just rendering and returning the value, the MetalROS now
saves the rendered string back it its self. Then the next time the
value is retrieved, it can skip being rendered.

This is done by defining a method that returns the rendered value.
This means using hash indexing method, `[]()`, will cause a rerender.

This doesn't make a huge difference when rendering a single file, but
should represent a big improvement when rendering multiple.

Also as the MetalROS is being traversed, it converts sub Hash's into
MetalROS. These sub MetalROS are also cached in the same way.