Add logging to Data.load method
A major source of lag in Metalware is extraneous file reads, in
particular the answer file.

This commit adds logging to the Data.load method so the file reads
can be tracked.

As the logger can not work before a config has been initialized, the
logging needs to be switched off whilst loading the config file.

Also a Config object needs to be initialized for all the tests, this
is done in spec_helper.rb. A different config can still be used as this
will reset the logging.