Cache configure.yaml between validation
The configure.yaml file has the option to be cached by the Loader. The
cached file can then be passed to the answer vaildation so it does not
need to load the file itself.

At this point in time, the validator will load the question file if
one is not provided. This can be removed at a latter date.