Remove RepoConfigParser from DomainTemplateRender
The RepoConfigParser is being replaced with the new namespace. This
is the first step of the replacement. Config parameters can now be
directly queried from the namespace and thus the config doesn't need
to be rendered seperately.

The Templater class methods have been updated to except the Alces
namespace instead of a Config. The RenderDomainTemplates tests have
been updated to match the new templater.

One test in the SharedExamples can not be fixed until all of Metalware
uses the new templater.