Switch Named to use new Alces namespace
Named is now initialized with the Alces namespace and nolonger uses
RepoConfigParser.

The metalware config doesn't need to be passed in as it can be retreived
from the Alces namespace.

The Templater methods have been adjusted so a dynamic namespace can be
included at the top level. This is required so the named namespace
can be included for the particular template being rendered.

The Alces namespace now automatically turns the dynamic namespace into
a MetalRecursiveOpenStruct so it can be navigated the same way as the
configs.