Update Templater spec to use the new namespace
The Templater tests have been updated to use the new Namespace. Some
of the test have been deleted as they nolonger are part of the spec
for the Templater.

A few tests are being skipped as they either need to be moved to a
different spec OR are currently broken. They will be fixed in future
commits.