Switch Each command to use the new Namespace
The Nodes object doesn't need to be used anymore to use the templater.
Instead a render_erb_template method has been included into all
Namespaces that use a HashMerger. This way it is easy to render for
a node.

The alces namespace needed to be adjusted as the config method was
renamed to metal_config