Create node_identifier mixin for commands
A few different commands implement a node_identifier that can switch
between a single node and a group of nodes.

To simplify the use of this switch with the new namespaces, a mixin
has been implemented. This way the code can be shared between the
seperate commands.

It assumes the node_identifier is always the first input and that the
options flag is specified as 'group'.

Also added a find_by_name method to the metal arrays as it will come in
useful in the future.

When prepended to a Command, it will provide an array of namespaces
that correspond to the cli input. PLEASE NOTE, the mixin must be prepended
instead of included. This is so the pre_setup required can occurr
automagically.