Migrate build files and build method
Currently the old `Node` class contains the build method for the node.
This commit migrates this to the `Node` namespace.

This is because the old `Node` is heavily built on `RepoConfigParser`
and `TemplatingConfiguration` which both load configs. However as the
configs are already loaded into the namespace, it is easier to migrate
the required methods then to adapt the old `Node` class.

Also mirgated the file merging methods from the old `Node` class. These
methods are better suited to HashMerger.