Migrate methods out of old Node
This commit migrates a bunch of methods out of Node. This is because
the old Node class has a bunch of unrelated methods making it hard to
replicate it's behaviour.

Instead it's methods have been moved to: FilePath, Commands::Build and
Namespaces::Node.

This caused a significant amount of changes in how the Build::Command
works. Therefore the Build command needed to be updated to match the
changes. The tests have also been updated. This required to updating
how the tests are mocked.