Meger group configs in reverse order
This insures the primary group (aka the first gender) config is merged
last.

Also removed the old Node class. The reverse order was the last thing
Node did that HashMerger did not.

Also removed hash merger tests from the old Node spec as they are now
handled by HashMerger. They will not be replicated as the current tests
are sufficient.