Remove legacy group index tests
The determination of a group index is a tricky task due to the
GroupCache. For this reason, the Group namespaces needed to be
initialized with a index and thus do not need a copy of the GroupCache

However as the GroupCache is already independantly tested, their is
duplicatation in the testing.

The additional tests of what to do if the Group does not appear in
the cache are now obsolete. It is not possible to create a Group
namespace for a node not in the Cache.