Rename the self node to local node
Due to the new namespaces ability to dynamically define method names
based on the Node name, having a node called self is problematic.

This is because self is a key word and would be dynamically overriden
possibile intoducing a bug. Instead the self node has been renamed
the local node.

The build_method will be updated in the next commit.