Create Local namespace and set it's build method
The local node has a dedicated namespace that is inherited from Node.
It is intended that the local node will behave similarly to a regular
Node.

The Node object now is made using `create` instead of `new`. The create
method will automatically use the `Local` namespace for the local node
(based on the name input).

The major difference between `Node` and `Local` is the build method.
The `Local` node will always use the local build. Where a regular
`Node` will error if it tries to use the local build.