Migrate Templater node url tests to new Nope spec
The majority of the tests are redundent as the files have already
been tested seperately in HashMerger. The build complete and kickstart
URL tests have been ported

Also the tests use to check the behaviour if a Node object was
initialized without a name. However this is no longer possible and
thus will not be tested.

Also removed the test that allows a node to be intialized that is
not in the genders file. If a node isn't in the genders file, it will
not appear in alces.nodes.