Migrate Tempater tests to Domain spec
The tests for the domain namespace have been migrated to their own
file.

The tests for all the genders have been removed as this feature is
not going to be implemented at this time. It can be added again in
the future if required.

Also removed the test for the hunter cache as they should be handled
by the hunter command