Skip the build integration tests
There is a bug in the build integration tests that is breaking the
`nodeattr` commands. This bug is somehow interacting with the other
tests.

As a result, the standard tests are failing only when the integration
tests are ran. Otherwise they are passing.

The integration test still need some work to get them working with
the new metalware and thus might be worth better mocking of the tests.