Create AlcesUtils to make mocking easier
With the new namespace, it can be directly mocked to make testing
easier. This commit creates the utility and applies it to Alces spec.

The utility creates a namespace that will be global within the test.
It allows the namespace to be mocked directly without having to
create config files.