Default to domain scope when rendering templates
If a scope is not specified for a template (by passing a config into
the dynamic namespace) it defaults to the domain.

A option in the config to switch it off has been included as some of
the tests assumed this wouldn't happen.

This can be fixed at a latter date.