Do not tell BuildMethod how to build
The build command should not be telling the build methods how to build
a node. The reponsibility of the build command is now to control the
flow of the build.

What gets rendered with what parameters is the responsibility of the
BuildMethod. Default options for the build parameters are now used.
However this isn't being enforced as not all situations may use them.

The comment about `render_build_start_templates` (note the new name),
has been removed as it no longer applies.

Also `render_build_complete_templates` does not to raise an error by
default as not all build methods need it.