Test that an error occurs for unset parameters
As the templater has got more complex, it is no longer reasonable
to convert all possible errors into UnsetParameterAccessError.

Instead the error type and back trace is retained, but the file that
failed to render is prepended to the message.

The test doesn't need to be rewritten otherwise as the Templater
class methods have been retained.