Create the NilDetectionWrapper and spec
The MissingParameterWrapper is going to be replaced by a new wrapper
class. It will be mostly the same but will also have tests that it is
working correctly.

Currently it can only detect if a single method call returns a nil.
It does not work for chain method calls currently.