Wrap alces with nil detection when templating
Nil detection will automatically be included when templating

Internally within metalware, alces can return a nil without a warning
if it does not go through the templater.

However this does not apply when templating as mentioned above.

ERB does not play well with the wrapper. It expects to_str to return
a String and not the wrapper. A condition has been added to detect
this and returnt he string.