Create utility for making mocked nodes
The mock_node method can be used to make a node without having to
set up configs, answers and gender files.

By default it is placed in the 'test-group' unless there is a group
input saying otherwise. It does not create the group, this would have
to be done seperately.

It does however add the node to the alces.nodes array. It maintains
the old nodes list so it is possible to add two nodes with the same
name. By default, the local_only genders file is used. So only the
local node should appear in the list.

It also sets the node to be `alces.node`. Please note, the last
mocked node is set to `alces.node`.

All mocked nodes have empty config and answer.