Use HASH_MERGER_DATA_STRUCTURE when mocking configs
The correct data structure is now being used when mocking configs.
However it can not render configs. The purpose of the mock is to
return predefined values.

The templater is fairly well tested as is. Therefore the higher order
tests that rely on particular config values shouldn't need use the
templater. Tests that need a mocked config should have a predefined
return value.