Remove unfixable test for build spec
The test previosuly checked it only rendered build files it could
find. However this is not strictly true as other files will be rendered
(such as kickstart/ pxelinux). This is what is causing the test to
fail. They should be ignored by the expect regex, but the exact
reason can not be determined.

A better way would be to check the rendered files. However due to
the mocking, the Templater does not run. It is not possible to make
the Templater run without making adjustments to the other tests.

In reality these are minor tests as the templater will fail to find
the files if the do not exist.