Remove genders input into the mock method
The genders file input was original used to ensure the local node
exists in the genders file. However rerunning the mock each time
could lead to some odd bugs.

Instead the genders file is mocked once at the start of each test
to include the local node. After that it is up to the user to ensure
each mock genders file includes the local node.