Create hasher mixin for the namespaces
The previous ObjectFieldHasifier works well however isn't fit for
hashing the namespaces. The Namespaces have public templating methods
that can't be hashed.

Instead a mixin has been created to do it. It uses two white lists
to specifiy the methods to be ran. The difference is one runs to_h
on the return value (recursive) and the other does not.