Implement the hash white list in HashMergerNamespace
The domain, group and node namespaces are now using a share example
test for HashMergerNamespace. This is so the new to_h method can
be tested.

By default, the answers and config will be hashed for the above
namespaces. The name will also be hashed if it is set. The code is
included in the base class so the mixin doesn't need to overide the
method

The AlcesUtils mock_group method sets the new group as the in scope
group.