Make note about DomainTemplatesRenderer
Currently this file is to large and needs to be refactorered.

Some of the managed file methods can be broken out into Templater.
Currently one of the Templater methods as a magic managed file parameter.
This should be replaced with the managed methods in DTR.

Also the validation render method code (in DTR) can be broken out
into their own classes. A similar approach to `BuildMethods` coule be
used.