Switch view-config to new view command
Instead of only viewing the config for a particular node, the view
command has been expand to show any value in the Namespace. It uses
the new AlcesCommand mixin for it's cli.

Therefore the only thing it needs to do is convert the cli object
into a hash and then JSON