Create AlcesCommand helper for cli inputs
There is now a requirement to take a string from the cli and convert
that into a namespace. This is going to be used by the view commands.

There are a few helper features in which `alces` can optionally
included at the begining of the string.

Also 'domain', 'groups', 'nodes' and 'local' do not need to be typed
in full. Please not that alces will not respond to `node` as it is
not in scope. So it means `node` will be interpreted as `nodes`.