Display the full hash for Node and Group
Both the Node and Group namespaces needed additional methods added
to there white lists. Now all the required fields are being included
in the hash.

The view command needed to be adjust so it does not try to convert
MetalArrays into hashes. Instead display the array directly.

When converting named namespaces into strings, the name is included.