Mock the nodeattr command from FakeFS genders
Instead of mocking the namespace objects when creating a node, it is
easier to directly update the genders file.

The problem with the genders file is it can not be used with the faked
filesystem. So instead there is mocking around the NodeattrInterface.

From now on, AlcesUtils automatically mocks the NodeattrInterface
to use a temporary file copy of the faked genders file. A new temporary
file is created for each nodeattr command so it does not become stale.

This way the faked genders file can be used as if it was the real file.