Fix broken tests from mocking genders in AlcesUtils
The templater spec now needs to use the faked filesystem, otherwise
no other change was required. The use of the real filesystem is legacy
code that is no longer required.

The render spec use to overide the File.read method, however this is
not compatiable with the new mocking of genders. Instead a temp file
is used.