Load the genders file into faked filesystem
When using AlcesUtils, the old SpecUtils.use_mock_genders method no
longer works. This is because the command to be used is passed into
the nodeattr method directly instead of using the Constants.

Instead, the genders file can be spoofed in the fake filesystem which
will be used by the nodeattr command.