Fix broken Node and Group to_h tests
The addition of the white list hasher to Node has caused the tests
to break. A additional utility to mock the hex ip of a node needed
to be added to AlcesUtils.

Also the group that the node belongs to needed to be created as well.
To do this, the default group name needs to be publicly availiable.

Also a seperate context has been added to the Node spec for with and
without AlcesUtils.

Also as the nodes list in Group and the group in Node are displayed
by to_h both need to exist. This means both sets of tests need to
mock both a group and a node.