Update AnswerTableCreator to use new namespace
The answer table creator needs to pull the answers from the new
namespace objects so it can take advantage of the cache. For this
reason the alces object must be passed in.

The tests no longer mocking the genders file nor write the answers
to a file.

Instead the group and node are mocked directly using AlcesUtils.
This means the filesystem setup is no longer required

The view-answer command has been updated so it uses the updated
AnswerTableCreator. However it isn't possible to render the answers
reliable as `alces.node` is not inscope for domain and group. However
as `alces.node` is commonly used at the domain and group levels, it
causes an error.