Make all the tests use the faked filesystem
Metalware is at the point in which we no longer should use the real
filesystem in the tests. Instead any required file should be cloned
into the faked filesystem.

This change is required as travis is failing. Metalware does not run
as sudo on travis and does not have permisions to create the Log file.
Instead of mocking the logging, the proplem can be fixed permanently
by using FakeFS