Add methods to modify the root faked fs
Now all the tests are running in a faked file system, it is useful
to configure it directly. Thus a root FileSystemConfigurator is used
by the tests. It can be configured with the root_setup method and will
be used automatically.

This removes the need for explicitly calling the configured file
system.