Obey the path given to nodeattr when mocking
In some cases, a gender file is given to nodeattr for validation.
In these cases, the given file (in the fake filesystem) needs to be
cloned to the real filesystem.

So by stripping the filepath from the command, the passed file can
be written to the temporary file in the real system.

This fixes the NodeattrInterface spec.