Test the each command with FakeFS on by default
Now that FakeFS is on by default, the each command needs to be mocked
dirrently. This is because the each command runs a shell command that
can not be mocked.

However the ruby process has to run within FakeFS. The solution to this
is to run the ruby command with FakeFS on but otherwise do the setup
with FakeFS off. This ensures the temporary file is in the real
filesystem and thus the bash command can write to it.