Run ruby NodeattrInterface code in faked FileSystem
As the ruby component of the code in NodeattrInterface logs the system
command, it must occur in the faked filesystem.

However as the bash command itself can not be mocked, the temporary
file must be in the real filesystem.

This commit makes the split between the two parts. It also ensures that
the temporary file is unlinked. Note that this commit is only required
by travis.