Move local answers into nodes folder
Up until this point, loading the local answers using the local method
is different to loading a node answer for the local node.

This meant that different files where loaded depending which method
was used. From now on, the local answers are in the nodes answer
directory.

In addition to this, a check has to be done when loading node answers
for the local node. This is because the local node is validated
differently to regular nodes. The check ensures the correct validation
is performed.