Fix order in which defaults and answers are merged
There is a bug in the order answer files and defaults are merged.
Currently the answer defaults are merged into the answer hash on
file load (after validation).

This causes a load order of:
domain_default, domain_answer, group_default, group_answer, node_...

The correct order (from 2017.1.0) should be:
default_domain, d_group, d_node, answer_domain, a_group, a_node

This is achieved by loading the defaults into an array and appending
it to the beginning of the answers. Also questions without a default
are ignored instead of loading a nil default.