Create benchmark for the build process
The benchmarking script will run the build command for a group.
The group is always 'testgroup' so the benchmarking will be differnt
depending on the configuration.

Inorder to end the benchmark, the nodes must be bilt. This can be
mocked with a seperate each command that touches the build complete
paths.

However this required the build_complete_path to be moved from local
to build.