Fix build files spec so they replace higher configs
When merging the build files in the configs, they should replace the
higher level. Currently they are appending, which is a change from
2017.1.0

Also made the files hash optional which is useful for testing. However
it also provides greater flexibility in the types of builds metalware
can perform.

The cached data is also deep duplicated to prevent the cache being
altered.