Log when a new alces namespace is made
The creation of a new namespace will not share the existing cache.
This allows for metalware to change the state of the cluster and
create a updated templater to match.

However it will also slow Metalware down as the files need to be
reloaded and validated. Therefore it is useful to log when this
happens.

Also removed the unnecessary Alces.new in build benchmark