Create render_to_staging method for templates
The intention is to render all templates to a staging directory before
syncing them into the file system. As such the render_to_staging method
will need to replicate the functionality of render_to_file. (This is
still yet to be fully implemented).

However it will always place the file in the staging directory. The
final save location is still required for the syncing step.