Implement render_to_file as an instance method
The previous method was defined at the class level. However it is
going to be deprecated as managed files will be handled by sync and
not at the templating stage.

However there may still be a use case where files need to be directly
rendered. For this reason it has been defined at the instance level
in a slightly modified form. This means it can also be used by the
render_to_staging method.