Do not validate staging files on render
The staging files can not be validated when they are rendered. This
is because managed files will only be merged during the sync and thus
can not be validated before that.

Instead a validation class name can be included as a flag. This will
be stored in the staging manifest and will instruct the sync command
on how to validate the file.