Render build files to staging with template cmd
A new template command has been implemented to render files to staging.
This is to replace the first step of the build command. Build will no
longer render templates (to be implemented *note exceptions tba).

The build methods needs to be adjusted to match this change. A new
render_to_staging method (in BuildMethod )has been created to handle
rendering to staging.

In Templater, the render_to_staging method needs to create the directory
where the file will be saved. This is because a file could be rendered
anywhere in the real filesystem which needs to be mirrored under staging.

Additional logging has been added to Data.