Implement Sync command that moves staging files
Once a staging file is rendered it needs to be moved into place. This
is completed by the first stage of the sync command. Currently only
non-managed files can be moved. Once a file has been moved, it is
removed from the manifest which is updated at the end.

Eventually sync will also handle service management.