Remove staging specific code from Templater
The Templater class no longer contains code that is used to handle
the Staging files. Instead the Templater class is initialised with
a staging object that is then used for the render.

This way the staging file handling and the rendering can be decoupled.

The 'update' method on Staging will always save the current version
of the manifest. This means multiple actions can be performed on it
without reloading the file. But then it will always be saved to
prevent it going stale.

There is a 'template' method on Staging which automatically creates a
Templater object. By doing so, all the rendering can share the same
staging cache which removes the need to reload the staging file. As
the template method is built on `update`, it will automatically save
the manifest at the end.

As a instance of Templater is now exclusively used for staging files,
the other render_to_file methods have been removed. Also staging files
are validated on sync and thus the validation block does not needs to
be ran.