Allow managed files to be synced from Staging
From now on, all managed files will be merged together on Sync. This
means the when rendering a managed file, only the rendered section is
saved in the staging directory. This is why validation must run on sync
and not when they are templated.

To accommodate this, this managed file code has been removed from the
Templater class and moved into its own class. This way the file handling
and merging of ManagedFile's can be decoupled.