Allow Staging to create directories on sync
In some cases, the sync directory needs to be made before the file
can be put into place. As it is not desirable for Metalware to do this
in all case, a mkdir flag has been included in staging.

Also Templater.new.render method now accepts key word arguments which
will be passed to Staging.push_file. This means it can handle all types
of staging files (including managed) without requiring a separate
methods for all cases.