Switch the BuildMethods to use hooks
This is the first commit to refactor the build process to be based
on hooks. The old render methods are no longer required as they have
been replaced by staging templates.

The only template that needs to be rendered at build are pxelinux. The
build process has been simplified down to two hooks (currently). A
start_hook that runs at the beginning. This is when the local node script
is ran and pxelinux firstboot is rendered.

Then there is a complete_hook that is used to rerender the pxelinux with
firstboot equal false. Only the pxelinux templates are rendered with
firstboot. This is because it isn't cacheable and must be handled
differently. Thus firstboot does not appear in the namespace by default.