Initialize the BuildMethod within the Node namespace
Instead of saving the class of the BuildMethod within the Node
namespace, it is going to be initialized. This is to simplify the
code internally within metalware. As the BuildMethod is an internal
metalware tool, this shouldn't affect the repo.

The BuildMethod class can still be found using node.build_method.class

This change requires the metal_config. To facilitate this, the config
is now a public method on alces. This way it does not need to be passed
around inorder to be used