Intermediary commit for Build refactor
Currently the build command is broken due to the changes to how the
build files are rendered and the new hooks. This commit starts the
refactor to fix the issues.

Build now assumes the build_method is an initialized class and thus
easier to mock in the tests. The build tests now include AlcesUtils to
do the mocking. This allows an automatic mapping of the genders file
from the Faked filesystem to the real one.

Fixed a bug in AlcesUtils where it wasn't deleting the temporary mocked
genders files it was making. This is because FakeFS needs to be turned
off in order to delete the file.