Fix build command tests and improve reporting
The error reporting for threads and the utilities has been improved.
Now if an error has occurred, it is caught and explicitly logged
before being raised again. This is because the threads have a habit
of making the error messages disappear.

The other major change is the build test. Previously build was
responsible for rendering the files. However now it runs BuidMethod
hooks instead. (NOTE: the Kickstart builds will render the pxelinux
template on start_hook).

This means it is easier to test if the hook has been called instead
of rendering the file.

There was a small bug in the pxelinux build method where the config
add been renamed to metal_config.