Remove support for changing the config via the cli
This functionality is no longer being used in testing and has never
been used in production. Instead, the user should change the default
config file contained within metalware.

The config still needs to be initialized at run time as it will change
with the --quite and --strict options from the cli. However as this
is not file paths, it does make a huge difference.

The CLI needs to load the config in order to get the default options.
This is achieved by using a cache that is set in the BaseCommand. If the
cache is accessed before it is set, it will error by default. However the
new_if_missing flag will return a new config if it isn't otherwise been
set. This means MetalLog will always work as it can always get a new config
if it hasn't already been cached.

MetalLog has been updated so it reads the Config cache and thus no longer
needs to be reset when the config changes. MetaLog was previously
causing problems as it needed to log both before and after the config
has been cached. Now it will automatically change once the cache has
been set.