Fix Config loading and remove skip_log from Data
As MetalLog retrieves the config from Config.cache, it can work before
the BaseCommand is called. This means that skip_log is no longer
required in Data.

However this means Data can not be used to load the config file. This
would create a recursive loop between Config, Data, and MetalLog.
The solution is to load the config file directly with YAML.