Fix broken test and add additional mock methods
As the Config only gets cached in the BaseCommand, it has caused most
of the tests to break. This is because the other classes are not allowed
to implicitly create a Config.

The cache is now being set the beginning of each test that has AlcesUtils
included. This is why the BaseCommand needs to clear the cache before
setting it.

The render_domain_templares tests have been skipped as they are now
broken. However as the DomainTemplateRenderer class is going to be
removed, they will not need to be fixed.

A new utility has been added to AlcesUtils to make running code with
redirected stdout, and stderr easier. This sort of code is used in a
few places, and thus should make handling it easier.