Create RenderMethods and implement Genders rendering
The genders file is to be rendered using a new RenderMethod class. This
is to abstract away the details on how a file is rendered and then
validated. This is because the files need to go through the template,
staging, sync cycle.

The RenderMethod contains the method which will render_to_staging. The
namespace to be used needs to be provided as well as the templater. The
reason the templater is an input, is so the staging_manifest does not
need to be reloaded for each file. Instead it is cached between templating.

The validate method then takes the content of a file. As the managed
files are merged in memory, only the file content is available. As
some validation requires system commands on files, a run_in_temp_file
method can be used.

The NodeattrInterface now errors if the validation fails. This will be
caught as part of the Sync command.