Stage the genders file on Configure and fix tests
The NodeattrInterface spec needed to be updated to reflect the fact it
no longer raises errors.

FilePath has been updated as the template_path method requires a node,
if the node is nil then it returns an error. Note that the node input
should be changed to a namespace input.

The Staging class methods have been updated to use the Config.cache

With the changes above, the genders file is now rendered to staging on
configure.